requirements.txt
PyYAML==5.3.1
requests==2.24.0
flask==1.1.2
flask-cors==3.0.9
redis==3.5.3